move maker to component area
